+++ src/cards/admin.py
@@ -1,5 +1,4 @@
 from django.contrib import admin
-from .models import Card, UserCard
+from .models import Card
 
 admin.site.register(Card)
-admin.site.register(UserCard)
